Compatibilidade com Arch

+++ run_once_10-apt.sh
@@ -1,21 +1,29 @@
 #!/bin/bash
 
-sudo apt update
-sudo apt install -y \
-  age \
-  bat \
-  btop \
-  build-essential \
-  curl \
-  fontconfig \
-  gcc \
-  git \
-  neofetch \
-  procps \
-  ripgrep \
-  tldr \
-  tmux \
-  unzip \
-  wget \
-  zoxide \
-  zsh
+# Check if the OS is Arch Linux
+if [[ -f /etc/arch-release ]]; then
+    echo "Using Arch"
+    sudo pacman -Sy git zsh
+else
+    # Update package list and install packages for Debian-based systems
+    sudo apt update
+    sudo apt install -y \
+      age \
+      bat \
+      btop \
+      build-essential \
+      curl \
+      fontconfig \
+      gcc \
+      git \
+      neofetch \
+      procps \
+      ripgrep \
+      tldr \
+      tmux \
+      unzip \
+      wget \
+      zoxide \
+      zsh
+fi
+
